First naive attempt at minimizing the execution time
This commit reduces the execution time on the data I have to about 35%:
From 1060s to 367s on my computer. Memory use should also be lower.

The aim of this change is to limit the copying of arrays and to use
linked lists (for O(1) insertion and deletion) to store nodes that are
in transitionary states as efficiently as possible.