Publish the profiling data I've kept for optimization
Keep in mind that this data can't be called benchmarks, it's just a single
cPython run on a workstation that's doing other things in the background.
It's mainly helpful to see major improvements.

The simulation's running time went down from 1060s to run a single time
with a timestep of 100ns, to taking 365 seconds in the current HEAD. The
simulation run with the default settings (with the timestep up to 1ms)
now takes 133 seconds to complete a single time, and 186s to complete
12 times on 4 different jobs.

Right now the program's biggest bottleneck is the adjacency_matrix
function, taking about 100s of the program's total running time on
default settings. The next major improvement in these profiles will be
by generating an artificial pore network as default behaviour.