Use same logic for adjancency_matrix and extract_throat_list
It is hard to tell why the networks are different and which one is right unless looking manually into the matrix.  I think the extract_throat_list was overly complicated because it was initially made to get the throat/connection properties. It rose my suspicion that the corresponding adjacency matrix was not symmetric and therefore I wrote the second routine.
I adapted the method, but I did not test it.